replaced some int occurrences

+++ Tracktivity/FinishActivityViewController.m
@@ -138,7 +138,7 @@
 	return cell;
 }
 
-- (int)tableView:(UITableView *)tableView numberOfRowsInSection:(NSInteger)section
+- (NSInteger)tableView:(UITableView *)tableView numberOfRowsInSection:(NSInteger)section
 {
 	switch (section) {
 		case 1:
